requirements.txt
Django>=2.2
celery==4.4.1
redis==3.4.1
psycopg2-binary==2.8.5
python-dotenv
django-celery-beat
django-cors-headers
djangorestframework
requests